# Trivia-Nexus/Trivia-Nexus/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quizara API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: Admin - list all categories
      responses:
        '200':
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Admin - create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryBody'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /admin/categories/{categoryId}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Admin - update a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryBody'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Admin - delete a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
  /admin/questions:
    get:
      operationId: adminListQuestions
      tags: [admin]
      summary: Admin - list all questions
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Questions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuestions'
    post:
      operationId: adminCreateQuestion
      tags: [admin]
      summary: Admin - create a new question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionBody'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /admin/questions/{questionId}:
    put:
      operationId: adminUpdateQuestion
      tags: [admin]
      summary: Admin - update a question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionBody'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
    delete:
      operationId: adminDeleteQuestion
      tags: [admin]
      summary: Admin - delete a question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
...
components:
...
  schemas:
...
    Category:
      type: object
      required: [id, name, nameAr, icon, parentId, questionCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        color:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
        parentId:
          type: [string, 'null']
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    CreateCategoryBody:
      type: object
      required: [name, icon]
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        color:
          type: string
        imageUrl:
          type: string
        parentId:
          type: [string, 'null']
    Question:
      type: object
      required: [id, categoryId, question, questionAr, options, correctAnswer, difficulty]
      properties:
        id:
          type: string
        categoryId:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, image, fill_blank, ordering, matching, audio, hotspot]
          default: multiple_choice
        question:
          type: string
        questionAr:
          type: string
        imageUrl:
          type: [string, 'null']
          description: URL of image for image-type questions
        options:
          type: array
          items:
            type: string
        optionsAr:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
          description: Index of the correct answer
        explanation:
          type: [string, 'null']
        explanationAr:
          type: [string, 'null']
        difficulty:
          type: integer
          minimum: 1
          maximum: 3
          description: 1=easy, 2=medium, 3=hard
    CreateQuestionBody:
      type: object
      required: [categoryId, question, options, correctAnswer]
      properties:
        categoryId:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, image, fill_blank, ordering, matching, audio, hotspot]
          default: multiple_choice
        question:
          type: string
        questionAr:
          type: string
        imageUrl:
          type: [string, 'null']
        options:
          type: array
          items:
            type: string
        optionsAr:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
          minimum: 0
        explanation:
          type: string
        explanationAr:
          type: string
        difficulty:
          type: integer
          minimum: 1
          maximum: 3
    PaginatedQuestions:
      type: object
      required: [questions, total, limit, offset]
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...